fix: change ThresholdStep type from int to float64

+++ panel.go
@@ -420,7 +420,7 @@ type (
 	}
 	ThresholdStep struct {
 		Color string `json:"color"`
-		Value *int   `json:"value"`
+		Value *float64   `json:"value"`
 	}
 	FieldConfigColor struct {
 		Mode       string `json:"mode"`
